add clusters, improve mapmouse events

+++ src/app/map/map.component.ts
@@ -1,11 +1,21 @@
 import { Component, DestroyRef, OnInit, inject } from '@angular/core';
 import { MapService } from '../services/map.service';
 import { EventsService } from '../services/events.service';
-import mapboxgl, { GeolocateControl, LngLat, Map, MapMouseEvent } from 'mapbox-gl';
+import mapboxgl, { GeolocateControl, LngLat, Map } from 'mapbox-gl';
 import { environment } from '../../environments/environment';
 import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
+
+const markerSymbolLayout: mapboxgl.SymbolLayerSpecification['layout'] = {
+  'icon-image': 'custom-marker',
+  'icon-anchor': 'bottom',
+  'text-font': ['Open Sans Bold', 'Arial Unicode MS Bold'],
+  'text-size': 14,
+  'text-offset': [0, 0.25],
+  'text-anchor': 'top',
+};
+
 @Component({
   selector: 'app-map',
   standalone: true,
@@ -28,9 +38,11 @@ export class MapComponent implements OnInit {
           lng: position.coords.longitude,
         };
         // set bounds for the autofill search
-        this.viewbox = `${position.coords.longitude - 0.05},${position.coords.latitude + 0.05
-          },${position.coords.longitude + 0.05},${position.coords.latitude - 0.05
-          }`;
+        this.viewbox = `${position.coords.longitude - 0.05},${
+          position.coords.latitude + 0.05
+        },${position.coords.longitude + 0.05},${
+          position.coords.latitude - 0.05
+        }`;
         this.eventService.setCurrentPosition(param);
         // update location field in create event form
         this.mapService.convertLatLngToAddress(param);
@@ -81,7 +93,10 @@ export class MapComponent implements OnInit {
       mapboxgl: mapboxgl as any,
     }).on('result', (event) => {
       const location = event.result;
-      const latLng = { lng: location.geometry.coordinates[0], lat: location.geometry.coordinates[1] } as LngLat;
+      const latLng = {
+        lng: location.geometry.coordinates[0],
+        lat: location.geometry.coordinates[1],
+      } as LngLat;
       this.createEventPopup(map, latLng);
       this.mapService.addTemporaryMarker(latLng);
       this.mapService.setSearchResult(location.place_name, latLng);
@@ -89,43 +104,40 @@ export class MapComponent implements OnInit {
   }
 
   initMarkers(): void {
-    const sub = this.eventService
-      .getEvents()
-      .subscribe((events) => {
-        const allEvents = events.joinedEvents.concat(events.otherEvents);
-        allEvents.forEach((event) => {
-          this.mapService.addMarker(event);
-        });
+    const sub = this.eventService.getEvents().subscribe((events) => {
+      const allEvents = events.joinedEvents.concat(events.otherEvents);
+      allEvents.forEach((event) => {
+        this.mapService.addMarker(event);
       });
-      this.destroyRef.onDestroy(()=>sub.unsubscribe())
+    });
+    this.destroyRef.onDestroy(() => sub.unsubscribe());
   }
 
   addListeners(map: Map) {
-    let markerClicked = false;
-    map.on('click', 'markers', (e: MapMouseEvent) => {
-      const id: string = e.features![0].properties!['id'];
-      markerClicked = true;
-      if (id) {
-        this.eventService.selectEventById(id);
-      } else {
-        //temporary marker selected, remove previous marker if any
-        this.mapService.removeTemporaryMarker();
-      }
-      setTimeout(() => {
-        markerClicked = false;
-      }, 1);
-    });
-
     map.on('click', (e) => {
-      setTimeout(() => {
-        //timeout so markerClicked flag can set if the other click event has been called
-        if (!markerClicked) {
-          this.mapService.addTemporaryMarker(e.lngLat);
-          // if true, there was no event selected -> the mouse is not on a marker
-          this.mapService.convertLatLngToAddress(e.lngLat);
-          this.createEventPopup(map, e.lngLat);
+      const features = map.queryRenderedFeatures(e.point, {
+        layers: ['markers', 'clusters'],
+      });
+      //Map clicked
+      if (features.length === 0) {
+        this.mapService.addTemporaryMarker(e.lngLat);
+        this.mapService.convertLatLngToAddress(e.lngLat);
+        this.createEventPopup(map, e.lngLat);
+      }
+      //Marker clicked
+      else if (features.length === 1) {
+        const id: string = features![0].properties!['id'];
+        if (id) {
+          this.eventService.selectEventById(id);
+        } else {
+          // Temporary marker selected, remove the previous marker if any
+          this.mapService.removeTemporaryMarker();
         }
-      }, 0);
+      }
+      //Cluster clicked (cluster click propagates to the marker as well, thats why length is 2)
+      else if (features.length === 2) {
+        map.flyTo({ center: e.lngLat, zoom: 16 });
+      }
     });
 
     // change cursor to pointer when hovering over a marker
@@ -150,7 +162,8 @@ export class MapComponent implements OnInit {
           if (error) throw error;
           //@ts-ignore
           map.addImage('custom-marker', image);
-        });
+        }
+      );
       this.addGeoJsonSource(map);
       this.addSymbolLayer(map);
       this.initMarkers();
@@ -167,7 +180,10 @@ export class MapComponent implements OnInit {
       data: geojsonData,
       cluster: true, // Enable clustering
       clusterRadius: 50, // Radius of each cluster when zoomed out
-      clusterMaxZoom: 14, // Max zoom level to cluster points
+      clusterMaxZoom: 14,
+      clusterProperties: {
+        markerCount: ['+', 1],
+      },
     });
     const source = map.getSource('event-markers') as mapboxgl.GeoJSONSource;
     this.mapService.setGeoJSONSource(source); // Set the GeoJSON source in MapService
@@ -175,21 +191,25 @@ export class MapComponent implements OnInit {
 
   addSymbolLayer(map: Map) {
     map.addLayer({
-      'id': 'markers',
-      'type': 'symbol',
-      'source': 'event-markers',
-      'layout': {
-        'icon-image': 'custom-marker',
-        'icon-anchor':'bottom',
-        // get the title name from the source's "title" property
+      id: 'markers',
+      type: 'symbol',
+      source: 'event-markers',
+      layout: {
+        'icon-allow-overlap': true,
+        'text-allow-overlap': true,
         'text-field': ['get', 'title'],
-        'text-font': [
-          'Open Sans Semibold',
-          'Arial Unicode MS Bold'
-        ],
-        'text-offset': [0, 0.25],
-        'text-anchor': 'top'
-      }
+        ...markerSymbolLayout,
+      },
+    });
+    map.addLayer({
+      id: 'clusters',
+      type: 'symbol',
+      source: 'event-markers',
+      filter: ['has', 'point_count'], // Only for clusters
+      layout: {
+        'text-field': '{markerCount} events',
+        ...markerSymbolLayout,
+      },
     });
   }
 
@@ -205,33 +225,38 @@ export class MapComponent implements OnInit {
       })
         .setLngLat(lngLat)
         .setOffset(-50)
-        .setHTML(`<span style="font-weight: bold; font-size:16px; cursor:pointer; color:white" id='popupBtn'">Create event</span>`)
+        .setHTML(
+          `<span style="font-weight: bold; font-size:16px; cursor:pointer; color:white" id='popupBtn'">Create event</span>`
+        )
         .addTo(map);
 
       this.stylePopup(popup);
 
-      document
-        .getElementById('popupBtn')!
-        .addEventListener('click', () => {
-          this.eventService.toggleEventForm();
-          document.getElementsByClassName('popup_w_btn')[0].remove();
-        });
+      document.getElementById('popupBtn')!.addEventListener('click', () => {
+        this.eventService.toggleEventForm();
+        document.getElementsByClassName('popup_w_btn')[0].remove();
+      });
     }
   }
   stylePopup(popup: mapboxgl.Popup): void {
-    const popupContent = popup.getElement()!.querySelector('.mapboxgl-popup-content')as HTMLElement;
-    const popupTip = popup.getElement()!.querySelector('.mapboxgl-popup-tip')as HTMLElement;
-    if (popupContent  && popupTip) {
+    const popupContent = popup
+      .getElement()!
+      .querySelector('.mapboxgl-popup-content') as HTMLElement;
+    const popupTip = popup
+      .getElement()!
+      .querySelector('.mapboxgl-popup-tip') as HTMLElement;
+    if (popupContent && popupTip) {
       popupContent.style.animation = 'bounce-in 0.5s ease';
-      popupTip.style.display ='none';
+      popupTip.style.display = 'none';
       popupContent.style.borderRadius = '16px';
       popupContent.style.opacity = '0';
-      popupContent.style.background = 'linear-gradient(135deg, rgb(207, 36, 75,0.9), rgba(232, 61, 88,0.9))';
+      popupContent.style.background =
+        'linear-gradient(135deg, rgb(207, 36, 75,0.9), rgba(232, 61, 88,0.9))';
       popupContent.style.boxShadow = '0 4px 8px rgba(0, 0, 0, 0.2)';
-   
+
       // Trigger the fade-in after a short delay (to allow rendering)
       setTimeout(() => {
-        popupContent.style.transition = 'opacity 0.5s ease-in-out'; 
+        popupContent.style.transition = 'opacity 0.5s ease-in-out';
         popupContent.style.opacity = '1';
       }, 100);
     }
